separated subsytemIDs back into their own constants class for quick readability
